Seperate _BaseUnits and move it to definitions.py
